pyproject.toml
[tool.poetry]
name = "async-bank-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.36"
databases = {extras = ["aiosqlite"], version = "^0.9.0"}
pydantic-settings = "^2.6.1"
alembic = "^1.14.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
